requirements.txt
streamlit
groqhuggingface_hub
requests
